# A lightweight server that does not require any third-party module