However it has more boilerplate code