@rproenca 2020-07-05
Apache AB results.
Command: `ab -c 10 -n 1000 -p sample.json -T application/json http://localhost:3000/`

Current code:
```
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 100 requests
Completed 200 requests
Completed 300 requests
Completed 400 requests
Completed 500 requests
Completed 600 requests
Completed 700 requests
Completed 800 requests
Completed 900 requests
Completed 1000 requests
Finished 1000 requests


Server Software:        
Server Hostname:        localhost
Server Port:            3000

Document Path:          /
Document Length:        2 bytes

Concurrency Level:      10
Time taken for tests:   52.958 seconds
Complete requests:      1000
Failed requests:        0
Total transferred:      201000 bytes
Total body sent:        188000
HTML transferred:       2000 bytes
Requests per second:    18.88 [#/sec] (mean)
Time per request:       529.583 [ms] (mean)
Time per request:       52.958 [ms] (mean, across all concurrent requests)
Transfer rate:          3.71 [Kbytes/sec] received
                        3.47 kb/s sent
                        7.17 kb/s total

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.2      0       5
Processing:   452  528 108.1    496    1512
Waiting:      452  527 108.0    495    1511
Total:        453  528 108.1    497    1513

Percentage of the requests served within a certain time (ms)
  50%    497
  66%    513
  75%    534
  80%    550
  90%    602
  95%    687
  98%    832
  99%   1044
 100%   1513 (longest request)
```

New code:
```
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 100 requests
Completed 200 requests
Completed 300 requests
Completed 400 requests
Completed 500 requests
Completed 600 requests
Completed 700 requests
Completed 800 requests
Completed 900 requests
Completed 1000 requests
Finished 1000 requests


Server Software:        
Server Hostname:        localhost
Server Port:            3000

Document Path:          /
Document Length:        29 bytes

Concurrency Level:      10
Time taken for tests:   52.069 seconds
Complete requests:      1000
Failed requests:        0
Total transferred:      151000 bytes
Total body sent:        188000
HTML transferred:       29000 bytes
Requests per second:    19.21 [#/sec] (mean)
Time per request:       520.693 [ms] (mean)
Time per request:       52.069 [ms] (mean, across all concurrent requests)
Transfer rate:          2.83 [Kbytes/sec] received
                        3.53 kb/s sent
                        6.36 kb/s total

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.2      0       3
Processing:   449  518  84.1    496    1607
Waiting:      449  517  83.9    495    1607
Total:        450  518  84.1    497    1607

Percentage of the requests served within a certain time (ms)
  50%    497
  66%    513
  75%    532
  80%    542
  90%    573
  95%    614
  98%    713
  99%    899
 100%   1607 (longest request)
```